Bug 1673553 part 104 - Change ICStub layout a bit more. r=iain
* Move the next_ field to ICCacheIRStub; fallback stubs don't have a next stub.
* Move the enteredCount_ field to the base class instead of duplicating in derived classes.
* Use uint8_t instead of uint32_t for small bit fields in ICState to shrink sizeof(ICState)
* Shrink Kind from uint16_t to uint8_t. Remove the magic-value bits to simplify the code and we don't have many bits available.

This gets all Baseline IC stubs down to 4 words on 64-bit:

1) code_
2) enteredCount_ + isFallback_ flag
3) next_ (CacheIR stubs) or icEntry_ (fallback stubs)
4) stubInfo_ (CacheIR stubs) or state_ + kind_ (fallback stubs)

static_asserts are added to document/check the size of fallback and CacheIR stubs.

This should be pretty optimal. The second word has some padding so gives us space
to add more data/flags to each ICStub in the future without increasing the size.

Differential Revision: https://phabricator.services.mozilla.com/D99465

UltraBlame original commit: 0a977402aec1cd3f5bc0cf9affd398f778053f72

+++ js/src/jit/ICState.h
@@ -35,10 +35,10 @@ class ICState {
   enum class Mode : uint8_t { Specialized = 0, Megamorphic, Generic };
 
  private:
-  uint32_t mode_ : 2;
+  uint8_t mode_ : 2;
 
   
-  uint32_t trialInliningState_ : 2;
+  uint8_t trialInliningState_ : 2;
